+++ src/components/scripts/python/pyenv_init.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+export PYENV_ROOT="$HOME/.pyenv"
+export PATH="$PYENV_ROOT/shims:$PATH"
+
+if which pyenv > /dev/null; then eval "$(pyenv init -)"; fi
+
+alias python="$(pyenv which python)"
+alias pip="$(pyenv which pip)"